Modified - refactor navigation_bottom_screen.dart
