+++ lib/core/screens/navigation_bottom_screen.dart
@@ -4,11 +4,11 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
-import '../classes/strings.dart';
-import '../provider/app_theme.dart';
-import '../widgets/main_drawer.dart';
-import '../widgets/profile_view.dart';
-import '../features/home_feature/presentation/home_screen.dart';
+import '../../classes/strings.dart';
+import '../../provider/app_theme.dart';
+import '../../widgets/main_drawer.dart';
+import '../../widgets/profile_view.dart';
+import '../../features/home_feature/presentation/home_screen.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class NavigationBottomScreen extends StatefulWidget {
